Feature(Components/TransactionBlock/TransactionBlock.jsx):
Created the new component that will interact directly between Incomes
and Expenses. It will have a summary on the front face, that will be
the sum of all the numbers added. For interacting with them, there will
be a button in the bottom, that will allow us to add, remove or edit
those elements. This feature is work in progress.